requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
imbalanced-learn==0.11.0
    # via imblearn
imblearn==0.0
    # via -r requirements.in
joblib==1.3.2
    # via
    #   imbalanced-learn
    #   scikit-learn
numpy==1.26.1
    # via
    #   -r requirements.in
    #   imbalanced-learn
    #   pandas
    #   scikit-learn
    #   scipy
pandas==2.1.2
    # via -r requirements.in
python-dateutil==2.8.2
    # via pandas
pytz==2023.3.post1
    # via pandas
pyyaml==6.0.1
    # via -r requirements.in
scikit-learn==1.3.2
    # via
    #   -r requirements.in
    #   imbalanced-learn
scipy==1.11.3
    # via
    #   imbalanced-learn
    #   scikit-learn
six==1.16.0
    # via python-dateutil
threadpoolctl==3.2.0
    # via
    #   imbalanced-learn
    #   scikit-learn
tzdata==2023.3
    # via pandas
